backend/requirements.txt
# Core dependencies for Manager Agent
litellm>=1.0.0
google-genai>=0.2.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Optional: For better error handling and utilities
python-dotenv>=1.0.0

# Optional: For Internet Agent web search capabilities
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0
langchain-community>=0.3.0

